requirements.txt
dash == 2.14.2
pandas == 2.2.3
plotly == 5.18.0
psycopg2-binary == 2.9.10
requests == 2.32.3
python-dotenv==1.0.1